requirements.txt
pynacl>=1.5.0
discord.py>=2.1.0
lxml>=4.9.2
plyvel>=1.5.0
python-dotenv>=0.21.0
pytz>=2022.7
requests>=2.31.0
yt_dlp>=2023.7.6
